Refactor Satella assets and improve UI/UX, now works on mobile
Rename and reorganize Satella JS/CSS assets (append _satella suffix). Update imports/HTML to point to index_satella.js, utilities_satella.js, particles-config_satella.js and satella.css; remove the old utilities-satella.js implementation. Behavior fixes: make the full overlay clickable to enter, keep center container off-screen initially, adjust credits fade timing, and improve audio handling (use #background-audio, handle autoplay promise, hide visible volume control on touch devices). JS improvements: switch hidden text class from .añadir-text to .add-text and update setTextRotation, streamline createParticles import, and minor cleanups/guards in createElement. CSS adjustments: prevent page scrolling/overflow, consolidate and reposition .flex-body and .musicplayer, fix .bio layout, add pull-down-menu styling and responsive media queries. Minor formatting tweak in particles config.
