+++ pages/satella/js/utilities_satella.js
@@ -1,10 +1,10 @@
 import "https://cdn.jsdelivr.net/npm/particlesjs@2.2.3/dist/particles.min.js"
-import * as particleconfigs from "./particles-config.js"
-// almacenar los elementos creados
+import * as particleconfigs from "./particles-config_satella.js"
+// store the created elements
 const elements = new Map();
 let currentLine = 0;
 let isTransitioning = false;
-  let showBioOrSocial = 1;
+let showBioOrSocial = 1;
   
 export function createElement({ tag = "div", container = null, name, ...props }) {
   const el = document.createElement(tag);
@@ -18,19 +18,19 @@ export function createElement({ tag = "div", container = null, name, ...props })
       continue;
     }
 
-    // guard 1: style como objeto
+    // guard 1: style as an object
     if (key === "style" && typeof value === "object") {
       Object.assign(el.style, value);
       continue;
     }
 
-    // guard 2: propiedad directa del elemento
+    // guard 2: direct property of the element
     if (key in el) {
       el[key] = value;
       continue;
     }
 
-    // guard 3: atributo HTML
+    // guard 3: HTML attribute
     el.setAttribute(key, value);
   }
 
@@ -45,8 +45,6 @@ export function createElement({ tag = "div", container = null, name, ...props })
   return el;
 }
 export function createParticles({ container, config, name }) {
-
-
   const particlesContainer = createElement({ id: name, className: name });
   container.appendChild(particlesContainer);
   particlesJS(name, particleconfigs[config]);
@@ -57,33 +55,53 @@ export function enterSite() {
   const credits = document.querySelector(".credits");
   const arrow=document.getElementById("arrow");
   startVideo()
-  //Animations
-  
+
+  // Animations
   overlay.className = `${overlay.className} fade-out fade-to-the-right`;
   
-  
   // Wait for the fade to finish (0.5 seconds) and then hide the overlay
+  // Start credits fade animation (if present) instead of hiding immediately
+  if (credits) {
+    credits.classList.add("fade-to-the-right");
+  }
+
   setTimeout(function () {
     overlay.style.display = "none";
-    credits.style.display = "none";
+    // Do not hide credits here; let its animation finish first
     setTextRotation();
-    changeElements();
-  }, 500); // Matches the duration of the transition, in milliseconds
+  }, 500); // Matches the duration of the overlay transition, in milliseconds
+
+  // After the credits' fade animation finishes, hide it to remove from layout.
+  // The CSS animation for .credits.fade-to-the-right is 2s; hide after 2s.
+  if (credits) {
+    setTimeout(function () {
+      credits.style.display = "none";
+    }, 2000);
+  }
 
 }
 function startVideo(){
   const video = document.querySelector(".background-video");
-  
-  const audio = document.querySelector(".musicplayer");
+  const audio = document.getElementById("background-audio");
+  const volumeSlider = document.getElementById("custom-volume-slider-satella");
+  const volumeContainer = volumeSlider ? volumeSlider.parentElement : null;
+
   if (video) video.play();
   if (audio) {
-    const defaultVolume=  navigator.maxTouchPoints > 0 ? 1 : .4;
+    let defaultVolume = 0.4;
+    if (navigator.maxTouchPoints > 0) {
+      // Hide the visible volume control container on touch devices, not the <audio> element itself
+      if (volumeContainer) volumeContainer.style.display = "none";
+      defaultVolume = 1;
+    }
     audio.volume = defaultVolume; // Set the volume to 0.4, all my friend say it's too loud and they will be deaf
-    audio.play();
+    // play() can return a promise that is rejected on some browsers if autoplay is blocked
+    audio.play().catch(() => {});
   }
- arrow.addEventListener("click",changeElements);
-}
 
+  const arrow = document.getElementById("arrow");
+  if (arrow) arrow.addEventListener("click", changeElements);
+}
 
 export function controlAudio() {
   const audio = document.getElementById("background-audio");
@@ -99,16 +117,13 @@ export function controlAudio() {
 }
 
 export function setTextRotation() {
-
-  const textLines = Array.from(document.querySelectorAll(".añadir-texto")).map(
+  const textLines = Array.from(document.querySelectorAll(".add-text")).map(
     (element) => element.innerText
   );
   const textElement = document.createElement("p");
   const imageAndTextContainer = document.querySelector(".username");
   changeLine({ textElement, imageAndTextContainer });
   console.log("entering");
-
-
 
   function changeLine({ textElement, imageAndTextContainer }) {
 
@@ -120,9 +135,7 @@ export function setTextRotation() {
     }
   }
 
-
   setInterval(function () {
-
     if (imageAndTextContainer && textElement) {
       // Remove the current text element
       imageAndTextContainer.removeChild(textElement);
@@ -134,9 +147,7 @@ export function setTextRotation() {
       changeLine({ textElement, imageAndTextContainer });
     }
   }, 10000); // 10 seconds to match with animation
-
 }
-
 
  function changeElements () {
     if (isTransitioning) return; 
@@ -169,4 +180,3 @@ export function setTextRotation() {
       showBioOrSocial = 0;
     }
   };
-
