+++ pages/satella/satella.css
@@ -3,6 +3,14 @@
  * Theme: Re:Zero-inspired layout.
  */
 
+/* Prevent unwanted scrolling on mobile and horizontal overflow */
+html, body {
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+  -webkit-overflow-scrolling: touch;
+}
+
 /* Click to enter */
 .click-to-enter {
   position: fixed;
@@ -59,21 +67,16 @@
 
 /* Body for Satella */
 body.satella {
-  margin: 0;
-  padding: 0;
-  height: 100vh;
-  overflow: hidden;
-}
-
-/* I created this so that the main flex is not in the body, which saves me headaches */
-.flex-body {
   display: flex;
   justify-content: center;
   align-items: flex-start;
   flex-wrap: wrap;
-  width: 100%;
-  height: 100%;
-  position: relative;
+  margin: 0;
+  padding: 0;
+  height: 100vh;
+  overflow: hidden; /* prevent page vertical scroll */
+  overscroll-behavior: contain; /* avoid overscroll bounce on mobile */
+  touch-action: manipulation;
 }
 
 .background-video {
@@ -153,23 +156,21 @@ img.satella {
 
 /* Bio */
 .flex-body-bio {
+  position: fixed;
+  inset: 0; 
   display: flex;
   justify-content: center;
   align-items: center;
-  flex-wrap: wrap;
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  top: 0vh;
 }
 
 .bio {
   background-color: rgba(0, 0, 0, 0.45);
-  padding-left: 40px;
-  padding-right: 40px;
+  padding: 24px 40px;
   border-radius: 20px;
   box-shadow: 0px 0px 10px 10px rgba(72, 40, 105, 0.45);
   animation: center 4.5s ease-in-out infinite;
+  max-width: 80vw;
+  width: min(700px, 90%);
 }
 
 .bio-list {
@@ -199,29 +200,28 @@ img.satella {
 }
 /* Bio End */
 
+/* Pull-down menu */
+.pull-down-menu {
+  position: absolute;
+  right: 10px;
+  top: 50%;
+  max-width: 120px;
+  cursor: pointer;
+  border-radius: 100px;
+  box-shadow: 0px 0px 10px 10px rgba(35, 0, 80, 0.45);
+  background-color: rgba(49, 0, 113, 0.5);
+}
+
 /* Music player */
 .musicplayer {
   position: fixed;
   bottom: 7%;
-  right: -2%;
+  right: 10px;
   width: max-content;
   border-radius: 100px;
   box-shadow: 0px 0px 10px 15px rgba(72, 40, 105, 0.45);
   transform: rotate(-90deg);
   transform-origin: center;
-  background-color: rgba(49, 0, 113, 0.5);
-}
-
-/* Pull-down menu */
-.pull-down-menu {
-  position: absolute;
-  right: -48px;
-  top: 50%;
-  transform: translateY(-50%);
-  max-width: 120px;
-  cursor: pointer;
-  border-radius: 100px;
-  box-shadow: 0px 0px 10px 10px rgba(35, 0, 80, 0.45);
   background-color: rgba(49, 0, 113, 0.5);
 }
 
@@ -367,3 +367,37 @@ img.satella {
 }
 
 /* End of Utilities */
+
+/* Media querys */
+
+@media screen and (max-width: 1000px) {
+  .pull-down-menu {
+    top: 68%;
+    right: 12px;
+    transform: translateY(-50%);
+    max-width: 90px;
+  }
+}
+
+@media screen and (max-width: 500px) {
+
+ .center {
+  width: 400px;
+ }
+
+ .img-satella {
+  margin-right: 0;
+ }
+
+ .bio {
+  width: 400px;
+  padding: 10px;
+ }
+
+ .pull-down-menu {
+    top: 70%;
+  }
+
+}
+
+/* End of Media querys */
